requirements.txt
# Sikkim AI Itinerary — backend requirements
# Install with: pip install -r requirements.txt

fastapi>=0.111.0,<0.120.0
uvicorn[standard]>=0.29.0,<0.35.0
pydantic>=2.7.0,<3.0.0
langgraph>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.4.0
langchain-groq>=0.1.6,<0.3.0

# Vector DB (swap in when graduating from in-memory filtering)
# chromadb>=0.5.0
# sentence-transformers>=3.0.0

# Local dev
python-dotenv>=1.0.0
